Refactor submission answers components
